Add journals for writes to be in-memory only
To allow this FUSE filesystem to leave the rootdir unmodified, all
writes must be buffered in-memory and never persisted. These writes
are stored in a linked list of journal structs. The read, getattr and
fgetattr calls also are updated to use the journals instead of relying
on the rootdir for everything.