Added struct file_size for truncate()
truncate() and ftruncate() modify a file's size. If the file shrinks,
any journal entries that are beyond the new size are removed and any
that have overlap with the new size are updated. If the file extends,
nothing is currently performed. In future, a journal entry for the
extended part should be created to hold null bytes ('\0'). To calculate
the file size for getattr() and fgetattr(), a new linked list of
file_size structs is added.

fsync() is redundant as there is nothing to flush to disk. All changes
are held in memory and must not touch the disk. Hence, this becomes a
no-op.