Make unlink() and rmdir() in-memory only
The file_size struct is updated to file_s. File type and whether its
been deleted `is_deleted` is added to the struct. Since unlinks are
mocked and do not reach the underlying filesystem, we need a way to
keep track of which files we have deleted.

This diff adds support for removing files, calls to unlink(), and
removing directories, calls to rmdir(), to the in-memory behaviour of
this filesystem.

However, this does not perform the checks that unlink() should perform
such as comparing file permissions against the user's. rmdir() checks
that the path it is trying to remove is a directory and that it's empty.